chore: Splitting admin settings
